+++ src/Service/Yaml/Linter/EmptyValueLinter.php
@@ -8,19 +8,19 @@ use Aeliot\Bundle\TransMaintain\Dto\LintYamlFilterDto;
 use Aeliot\Bundle\TransMaintain\Model\ReportBag;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\FileManipulator;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\FileMapFilter;
+use Aeliot\Bundle\TransMaintain\Service\Yaml\FileToSingleLevelArrayParser;
 
 final class EmptyValueLinter implements LinterInterface
 {
     use EmptyPresetsTrait;
-    use GlueKeysTrait;
 
-    private FileManipulator $fileManipulator;
     private FileMapFilter $fileMapFilter;
+    private FileToSingleLevelArrayParser $fileParser;
 
-    public function __construct(FileManipulator $fileManipulator, FileMapFilter $fileMapFilter)
+    public function __construct(FileMapFilter $fileMapFilter, FileToSingleLevelArrayParser $fileParser)
     {
-        $this->fileManipulator = $fileManipulator;
         $this->fileMapFilter = $fileMapFilter;
+        $this->fileParser = $fileParser;
     }
 
     public function getKey(): string
@@ -37,7 +37,7 @@ final class EmptyValueLinter implements LinterInterface
             $empty = [];
             foreach ($localesFiles as $locale => $files) {
                 foreach ($files as $file) {
-                    foreach ($this->glueKeys($this->fileManipulator->parse($file)) as $translationId => $value) {
+                    foreach ($this->fileParser->parse($file) as $translationId => $value) {
                         if ('' === trim($value)) {
                             if (!\array_key_exists($translationId, $empty)) {
                                 $empty[$translationId] = [];