+++ src/Service/Yaml/Linter/KeysDuplicatedLinter.php
@@ -6,20 +6,18 @@ namespace Aeliot\Bundle\TransMaintain\Service\Yaml\Linter;
 
 use Aeliot\Bundle\TransMaintain\Dto\LintYamlFilterDto;
 use Aeliot\Bundle\TransMaintain\Model\ReportBag;
-use Aeliot\Bundle\TransMaintain\Service\Yaml\FileManipulator;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\FileMapFilter;
+use Aeliot\Bundle\TransMaintain\Service\Yaml\FileToSingleLevelArrayParser;
 
 final class KeysDuplicatedLinter implements LinterInterface
 {
-    use GlueKeysTrait;
-
-    private FileManipulator $fileManipulator;
     private FileMapFilter $fileMapFilter;
+    private FileToSingleLevelArrayParser $fileParser;
 
-    public function __construct(FileManipulator $fileManipulator, FileMapFilter $fileMapFilter)
+    public function __construct(FileMapFilter $fileMapFilter, FileToSingleLevelArrayParser $fileParser)
     {
-        $this->fileManipulator = $fileManipulator;
         $this->fileMapFilter = $fileMapFilter;
+        $this->fileParser = $fileParser;
     }
 
     public function getKey(): string
@@ -42,7 +40,7 @@ final class KeysDuplicatedLinter implements LinterInterface
                 $values = [];
                 $duplicatedKeys = [];
                 foreach ($files as $file) {
-                    foreach ($this->glueKeys($this->fileManipulator->parse($file)) as $translationId => $value) {
+                    foreach ($this->fileParser->parse($file) as $translationId => $value) {
                         if (\array_key_exists($translationId, $values)) {
                             $duplicatedKeys[] = $translationId;
                         } else {