+++ src/Service/Yaml/Linter/SameValueLinter.php
@@ -6,21 +6,20 @@ namespace Aeliot\Bundle\TransMaintain\Service\Yaml\Linter;
 
 use Aeliot\Bundle\TransMaintain\Dto\LintYamlFilterDto;
 use Aeliot\Bundle\TransMaintain\Model\ReportBag;
-use Aeliot\Bundle\TransMaintain\Service\Yaml\FileManipulator;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\FileMapFilter;
+use Aeliot\Bundle\TransMaintain\Service\Yaml\FileToSingleLevelArrayParser;
 
 final class SameValueLinter implements LinterInterface
 {
     use EmptyPresetsTrait;
-    use GlueKeysTrait;
 
-    private FileManipulator $fileManipulator;
     private FileMapFilter $fileMapFilter;
+    private FileToSingleLevelArrayParser $fileParser;
 
-    public function __construct(FileManipulator $fileManipulator, FileMapFilter $fileMapFilter)
+    public function __construct(FileMapFilter $fileMapFilter, FileToSingleLevelArrayParser $fileParser)
     {
-        $this->fileManipulator = $fileManipulator;
         $this->fileMapFilter = $fileMapFilter;
+        $this->fileParser = $fileParser;
     }
 
     public function getKey(): string
@@ -37,34 +36,35 @@ final class SameValueLinter implements LinterInterface
             foreach ($localesFiles as $locale => $files) {
                 $same = [];
                 $values = [];
-                $pairs = array_merge(
-                    ...array_map(
-                        fn(string $x): array => iterator_to_array($this->glueKeys($this->fileManipulator->parse($x))),
-                        $files
-                    )
-                );
-                foreach ($pairs as $translationId => $translation) {
-                    if (!\array_key_exists($translation, $values)) {
-                        $values[$translation] = $translationId;
-                    } else {
-                        if (!\array_key_exists($translation, $same)) {
-                            $same[$translation] = [$values[$translation]];
-                        }
+                foreach ($files as $file) {
+                    foreach ($this->fileParser->parse($file) as $translationId => $translation) {
+                        if (!\array_key_exists($translation, $values)) {
+                            $values[$translation] = $translationId;
+                        } else {
+                            if (!\array_key_exists($translation, $same)) {
+                                $same[$translation] = [$values[$translation]];
+                            }
 
-                        $same[$translation][] = $translationId;
+                            $same[$translation][] = $translationId;
+                        }
                     }
                 }
 
-                ksort($same);
-
-                foreach ($same as $translation => $translationIds) {
-                    sort($translationIds);
-                    $bag->addLine($domain, $locale, $translation, $translationIds);
-                }
+                $this->addLines($bag, $domain, $locale, $same);
             }
         }
 
         return $bag;
+    }
+
+    private function addLines(ReportBag $bag, $domain, $locale, array $same): void
+    {
+        ksort($same);
+
+        foreach ($same as $translation => $translationIds) {
+            sort($translationIds);
+            $bag->addLine($domain, $locale, $translation, $translationIds);
+        }
     }
 
     private function createReportBag(): ReportBag